java/TobySpring/src/test/java/com/example/tobyspring/controller/HelloControllerTest.java
package com.example.tobyspring.controller;

import org.junit.jupiter.api.Test;

import static org.assertj.core.api.Assertions.assertThat;
import static org.assertj.core.api.Assertions.assertThatThrownBy;

class HelloControllerTest {

    @Test
    void helloControllerTest() {
        HelloController helloController = new HelloController(name -> name);
        String result = helloController.hello("test");
        assertThat(result).isEqualTo("test");
    }

    @Test
    void failHelloControllerTest() {
        HelloController helloController = new HelloController(name -> name);

        assertThatThrownBy(() -> helloController.hello(null))
                .isInstanceOf(IllegalArgumentException.class);

        assertThatThrownBy(() -> helloController.hello(""))
                .isInstanceOf(IllegalArgumentException.class);
    }
}